backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --allow-unsafe --index-url=https://pypi.org/simple --no-emit-index-url --output-file=requirements-tmp.txt requirements-packaging-tmp.in requirements.in setup.cfg
#
alembic==1.9.1
    # via opnreco (setup.cfg)
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
colander==2.0
    # via opnreco (setup.cfg)
defusedxml==0.7.1
    # via opnreco (setup.cfg)
greenlet==2.0.1
    # via sqlalchemy
hupper==1.11
    # via pyramid
idna==3.4
    # via requests
iso8601==1.1.0
    # via colander
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via mako
paste==3.5.2
    # via opnreco (setup.cfg)
pastedeploy==3.0.1
    # via plaster-pastedeploy
pip==22.3.1
    # via -r requirements-packaging-tmp.in
plaster==1.1.2
    # via
    #   plaster-pastedeploy
    #   pyramid
plaster-pastedeploy==1.0.1
    # via pyramid
psycopg2==2.9.5
    # via
    #   opnreco (setup.cfg)
    #   sqlalchemy
pyramid==2.0
    # via
    #   opnreco (setup.cfg)
    #   pyramid-retry
    #   pyramid-tm
pyramid-retry==2.1.1
    # via opnreco (setup.cfg)
pyramid-tm==2.5
    # via opnreco (setup.cfg)
python-dateutil==2.8.2
    # via opnreco (setup.cfg)
python-dotenv==0.21.0
    # via opnreco (setup.cfg)
pytz==2022.7
    # via opnreco (setup.cfg)
requests==2.28.1
    # via opnreco (setup.cfg)
setuptools==65.6.3
    # via
    #   -r requirements-packaging-tmp.in
    #   paste
    #   pyramid
    #   zope-deprecation
    #   zope-interface
    #   zope-sqlalchemy
six==1.16.0
    # via
    #   paste
    #   python-dateutil
sqlalchemy[postgresql]==1.4.46
    # via
    #   alembic
    #   opnreco (setup.cfg)
    #   zope-sqlalchemy
transaction==3.0.1
    # via
    #   opnreco (setup.cfg)
    #   pyramid-tm
    #   zope-sqlalchemy
translationstring==1.4
    # via
    #   colander
    #   pyramid
urllib3==1.26.13
    # via requests
venusian==3.0.0
    # via pyramid
waitress==2.1.2
    # via opnreco (setup.cfg)
watchdog==2.2.1
    # via opnreco (setup.cfg)
webob==1.8.7
    # via pyramid
wheel==0.38.4
    # via -r requirements-packaging-tmp.in
xlrd==1.2.0
    # via opnreco (setup.cfg)
zope-deprecation==4.4.0
    # via pyramid
zope-interface==5.5.2
    # via
    #   pyramid
    #   pyramid-retry
    #   transaction
    #   zope-sqlalchemy
zope-sqlalchemy==1.6
    # via opnreco (setup.cfg)
